Added scilinux, fixed centos based on learnings from that, tested all ubuntus.
Makefile.all: chroot for dev moved to scilinux, back to rpm-packages.  Fix glob for copying deb
  packages since it was getting a directory.  Make sure to run all the commands in the rpmbootstrap
  directory.  I'd made lots of changes but hadn't gotten them properly separately committed.

pb: log when we start and finish the build, that's an important state.  Add missing setting of
  all_ok to false that kept us from properly stopping on errors.  Fix typos in comments.  Greatly
  simplify pb_get_distros function by using split and join.  Also remove whitespace since
  multi-line conf file support will cause that to be added.

Filter.pm: use new pb_pbos_to_keylist function to generate the list of basenames we want, and use a
  loop rather than lots of separate statements.  Simplifies and makes more powerful this function.
  Also now guaranteed to maintain consistency with key lookups in the hash maps.